// apps/avto-shop/src/libs/enums/member.enum.ts
export enum Type {
    USER = 'USER',
    AGENT = 'AGENT',
    ADMIN = 'ADMIN',
    DEALER = 'DEALER',
    SERVICE = 'SERVICE',
    SELLER = 'SELLER',
};
registerEnumType(Type, {
    name: 'Type',
});

export enum Status {
    ACTIVE = 'ACTIVE',
    DELETE = 'DELETE',
    BLOCK = 'BLOCK',
};